+++ duality/CLIexe/__init__.py
@@ -1,0 +1,4 @@
+#This folder contains the executeable client with .exe extenstion.
+
+#dualityCLI.exe version.
+from duality.cli.cli import __CLIexeversion__