+++ duality/__main__.py
@@ -1,4 +1,4 @@
 from duality.cli.cli import *
 
 if __name__ == '__main__':
-    dualityCLI()
+    CLI()